TRECISION: use float math functions on floats

+++ engines/trecision/actor.cpp
@@ -373,8 +373,8 @@ void Actor::actorDoAction(int action) {
 	_vm->_pathFind->reset(0, px, pz, theta);
 
 	float t = ((270.0f - theta) * M_PI * 2) / 360.0f;
-	float ox = cos(t);
-	float oz = sin(t);
+	float ox = cosf(t);
+	float oz = sinf(t);
 
 	SVertex *v = _characterArea;
 	float firstFrame = frameCenter(v);
